Remove recursive flag from rm command
I would like to suggest removing the -r (recursive) flag as it is not needed since the arguments should only be files.  

It is a bad habit to default to appending the `-rf` flags to `rm`.  Especially in cases where users are encouraged to copy and paste.   

Apologies if this submission isn't the preferred method for minor documentation tweaks.  Please let me know if you prefer different channels to raise/suggest changes like this.